small rss sources refactoring + add initial rss sources
