--- src/main/kotlin/dataTypes/RssSource.kt
@@ -1,3 +1,0 @@
-package prod.prog.dataTypes
-
-data class RssSource(val sourceUrl: String)
