+++ src/main/kotlin/dataTypes/rss/RssSource.kt
@@ -1,0 +1,3 @@
+package prod.prog.dataTypes.rss
+
+sealed class RssSource(val sourceUrl: String)
